fix(tests): merge all tests into single file for sequential execution
Vitest was running test files in parallel despite configuration.
Merging environment, build, and runtime tests into a single file
guarantees sequential execution within nested describe blocks.
